+++ RobotGoFast/LinearFit.cpp
@@ -1,20 +1,20 @@
 #include "LinearFit.h"
 
-const double LinearFit::TOO_CLOSE = 0;
-const double LinearFit::TOO_FAR = 0xFFFF;
+const float LinearFit::TOO_CLOSE = 0;
+const float LinearFit::TOO_FAR = 0xFFFF;
 
 LinearFit::LinearFit(int pin) : pin(pin) {
   numDatapoints = 0;
   pinMode(pin, INPUT);
 }
 
-void LinearFit::addDatapoint(double x, double y) {
+void LinearFit::addDatapoint(float x, float y) {
   this->x[numDatapoints] = x;
   this->y[numDatapoints] = y;
   numDatapoints++;
 }
 
-double LinearFit::read() {
+float LinearFit::read() {
   
   int raw = analogRead(pin);
 
@@ -42,7 +42,7 @@ double LinearFit::read() {
   
 }
 
-double LinearFit::lerp(double x, double x0, double y0, double x1, double y1) {
+float LinearFit::lerp(float x, float x0, float y0, float x1, float y1) {
 
   return y0 + (x - x0) * ((y1 - y0) / (x1 - x0));
